# docs/swagger.yaml
definitions:
...
  handlers.UserInput:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  models.PriorityEnum:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
...
  models.StatusEnum:
    enum:
    - new
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - New
    - InProgress
    - Done
  models.Task:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/models.PriorityEnum'
      project_id:
        type: integer
      responsible_user_id:
        type: integer
      status:
        $ref: '#/definitions/models.StatusEnum'
      title:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      registration_date:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is project management service API
  title: Project Management Service API
paths:
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: No users found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
        "400":
          description: Missing required fields
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User deleted
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get user by id
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            type: string
        "400":
          description: Missing required fields
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update user
      tags:
      - users
  /users/{id}/tasks:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: No tasks found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get user tasks
      tags:
      - users
  /users/search:
    get:
      parameters:
      - description: User email
        in: query
        name: email
        type: string
      - description: User name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Missing email or name parameter
          schema:
            type: string
        "404":
          description: No users found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Search user
      tags:
      - users
swagger: "2.0"